refactor: move files to different directories
