+++ src/components/Expenses/Expenses.js
@@ -1,4 +1,4 @@
-import Card from './Card';
+import Card from '../UI/Card';
 import ExpenseItem from './ExpenseItem';
 import './Expenses.css';
